Obsługa zamówienia dla ekranu własna pizza
Autor: Daniel Blokus
Sprawdzał: Maciej Hamulak

Zmiany:
- Obsługa zamówienia dla kreatora własnej pizzy
- Dodanie metody przypisującej cenę pizzy własnej do jej rozmiaru
- Dodanie metody ograniczającej liczbę składników do wyboru
- Dodanie metody odpowiedzialnej za wyświetlenie komunikatu przy
wprowadzaniu niewłaściwych danych w ekranie własna pizza
- Dodanie metody odpowiedzialnej za wyświetlenie błędu przy wyborze
więcej niż 6 składników.
- Dodanie funkcjonalności do komponentów na ekranie własna pizza

+++ PWZP/src/app/Zamowienie.java
@@ -67,9 +67,10 @@ public class Zamowienie {
 	 * @param float1
 	 * @param dec
 	 */
-	void wyswietlLacznyKoszt(JLabel lbl, float float1, DecimalFormat dec){
+	void wyswietlLacznyKoszt(JLabel lbl, double float1, DecimalFormat dec){
 		lbl.setText((dec.format(float1)));
 	}
+
 	
 	/**
 	 * Metoda służąca do wyświetlania informacji: nazwa pizzy, rozmiar pizzy, liczba pizz na podglądzie zamówienia na ekranie dostawy
@@ -103,7 +104,6 @@ public class Zamowienie {
 		      exc.printStackTrace();
 		   }
 	}
-	
 	/**
 	 * Metoda służąca do wyświetlania wybranego sosu na podglądzie zamówienia na ekranie dostawy
 	 * @param s
@@ -236,5 +236,6 @@ public class Zamowienie {
 			exc.printStackTrace();
 		}
 	}
+	
 
 }
